Lowered spawn chances of boxes as rewards
Previously, a quest would in average hold around 15 chests (due to maxMarketValue being 5500) Lowering it to 1000 makes it so in average only 3 chests show up.
Note that the actual amount of chests that show up depends on their market value.
Their weight was also halved to lower the chances of them spawning over other loot.